// src/generated/AgentEnvironmentDiagnostic.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentEnvironmentDiagnosticKind } from "./AgentEnvironmentDiagnosticKind";
import type { AgentEnvironmentDiagnosticResource } from "./AgentEnvironmentDiagnosticResource";
import type { AgentResourceSource } from "./AgentResourceSource";

export type AgentEnvironmentDiagnostic = { resource: AgentEnvironmentDiagnosticResource, kind: AgentEnvironmentDiagnosticKind, message: string, path: string | null, source: string | null, sourceScope: AgentResourceSource | null, };