// src/generated/AgentEnvironmentSnapshot.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentEnvironmentDiagnostic } from "./AgentEnvironmentDiagnostic";
import type { AgentEnvironmentResources } from "./AgentEnvironmentResources";
import type { AgentModelRoute } from "./AgentModelRoute";
import type { ModelControlsSnapshot } from "./ModelControlsSnapshot";

export type AgentEnvironmentSnapshot = { modelControls: ModelControlsSnapshot, modelRoutes: Array<AgentModelRoute>, resources: AgentEnvironmentResources, diagnostics: Array<AgentEnvironmentDiagnostic>, };