// src/generated/AgentPackageSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentResourceSource } from "./AgentResourceSource";

export type AgentPackageSummary = { id: string, name: string, enabled: boolean, source: AgentResourceSource, filtered: boolean, installed: boolean, };