// src/generated/AgentResourceSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AgentResourceOrigin } from "./AgentResourceOrigin";
import type { AgentResourceSource } from "./AgentResourceSource";

export type AgentResourceSummary = { id: string, name: string, enabled: boolean, source: AgentResourceSource, origin: AgentResourceOrigin, };